# FSA.Tickets-Please/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SIMS incident/signals management API",
    "version": "v1"
  },
  "paths": {
    "/api/Addresses/Find": {
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Find address",
        "operationId": "FindAddress",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddressSearchModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SimsAddressViewModel"
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Addresses": {
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Add Address",
        "operationId": "AddAddress",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsAddressViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsAddressViewModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Addresses"
        ],
        "summary": "Update Address",
        "operationId": "UpdateAddress",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SimsAddressViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsAddressViewModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Addresses/{addressId}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Fetch address",
        "operationId": "GetAddress",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsAddressViewModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/api/Products/Addresses/{incidentSignal}/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Fetch product addresses",
        "description": "Fetch product addresses",
        "operationId": "GetProductAddresses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductFboAddressViewModel"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddressSearchModel": {
      "type": "object",
      "properties": {
        "search": {
          "type": "string"
        }
      }
    },
    "SimsAddressViewModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "mainContact": {
          "type": "string"
        },
        "addressLine1": {
          "type": "string"
        },
        "addressLine2": {
          "type": "string"
        },
        "townCity": {
          "type": "string"
        },
        "county": {
          "type": "string"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "postCode": {
          "type": "string"
        },
        "telephoneNumber": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "SimsFboTypes": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        4,
        8,
        16,
        32,
        64,
        128,
        256,
        512,
        1024,
        2048,
        4096,
        8192,
        16384,
        32768,
        65536
      ],
      "type": "integer"
    },
    "SimsAddressContact": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "telephoneNumber": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "isMain": {
          "type": "boolean"
        }
      }
    },
    "ProductFboAddressViewModel": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "fboTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsFboTypes"
          }
        },
        "mainContact": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "addressLine1": {
          "type": "string"
        },
        "addressLine2": {
          "type": "string"
        },
        "townCity": {
          "type": "string"
        },
        "county": {
          "type": "string"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "postCode": {
          "type": "string"
        },
        "telephoneNumber": {
          "type": "string"
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "contacts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsAddressContact"
          }
        }
      }
    },
...
  }
}